skip most forward pass steps once we reach the maximum amount of dsps

+++ openroad_interface/openroad_run.py
@@ -21,7 +21,7 @@ from . import scale_lef_files as scale_lef
 DIE_CORE_BUFFER_SIZE = 50
 
 class OpenRoadRun:
-    def __init__(self, cfg, codesign_root_dir, tmp_dir):
+    def __init__(self, cfg, codesign_root_dir, tmp_dir, run_openroad):
         """
         Initialize the OpenRoadRun with configuration and root directory.
 
@@ -32,6 +32,7 @@ class OpenRoadRun:
         self.codesign_root_dir = codesign_root_dir
         self.tmp_dir = tmp_dir
         self.directory = os.path.join(self.codesign_root_dir, f"{self.tmp_dir}/pd")
+        self.run_openroad = run_openroad
 
     def run(
     self,
@@ -121,15 +122,18 @@ class OpenRoadRun:
         """
 
         logger.info("Setting up environment for place and route.")
-        if os.path.exists(self.directory):
-            logger.info(f"Removing existing directory: {self.directory}")
-            shutil.rmtree(self.directory)
-        os.makedirs(self.directory)
-        logger.info(f"Created directory: {self.directory}")
-        shutil.copytree(os.path.dirname(os.path.abspath(__file__)) + "/tcl", self.directory + "/tcl")
-        logger.info(f"Copied tcl files to {self.directory}/tcl")
-        os.makedirs(self.directory + "/results")
-        logger.info(f"Created results directory: {self.directory}/results")
+        if self.run_openroad:
+            if os.path.exists(self.directory):
+                logger.info(f"Removing existing directory: {self.directory}")
+                shutil.rmtree(self.directory)
+            os.makedirs(self.directory)
+            logger.info(f"Created directory: {self.directory}")
+            shutil.copytree(os.path.dirname(os.path.abspath(__file__)) + "/tcl", self.directory + "/tcl")
+            logger.info(f"Copied tcl files to {self.directory}/tcl")
+            os.makedirs(self.directory + "/results")
+            logger.info(f"Created results directory: {self.directory}/results")
+        else:
+            logger.info("Skipping setup, using previous openroad results.")
 
         self.update_area_constraint(area_constraint)
 
@@ -359,7 +363,10 @@ class OpenRoadRun:
 
         # run openroad
         logger.info("Starting estimated place and route.")
-        self.run_openroad_executable()
+        if self.run_openroad:
+            self.run_openroad_executable()
+        else:
+            logger.info("Skipping openroad run, as resource constraints have been reached in a previous iteration.")
 
         wire_length_df = est.parse_route_guide_with_layer_breakdown(self.directory + "/results/codesign_codesign-tcl.route_guide")
         wire_length_by_edge = {}
@@ -447,7 +454,10 @@ class OpenRoadRun:
 
         # run openroad
         logger.info("Starting detailed place and route.")
-        self.run_openroad_executable()
+        if self.run_openroad:
+            self.run_openroad_executable()
+        else:
+            logger.info("Skipping openroad run, as resource constraints have been reached in a previous iteration.")
 
         # run parasitic_calc and length_calculations
         graph, _ = self.coord_scraping(graph, node_to_num)
